[fix] id type change (Integer->Long)

+++ src/main/java/com/example/seproject/issue/IssueRepository.java
@@ -1,6 +1,6 @@
 package com.example.seproject.issue;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface IssueRepository extends JpaRepository<Issue,Integer> {
+public interface IssueRepository extends JpaRepository<Issue,Long> {
 
 }
